Rename view_recarray -> recarray
