+++ submission/parsers/fr_log_parser.py
@@ -8,7 +8,7 @@ import pandas as pd
 
 from .base_log_parser import BaseSessionLogParser, UnknownExperimentError
 from .system2_log_parser import System2LogParser
-from ..viewers.view_recarray import strip_accents
+from ..viewers.recarray import strip_accents
 
 
 class FRSessionLogParser(BaseSessionLogParser):
