+++ submission/parsers/system3_log_parser.py
@@ -292,67 +292,3 @@ class Sys3EventsParser(BaseSys3LogParser):
                 matched_events.stim_params[field]=stim_event.stim_params[field]
             events[((events.stim_params.id == stim_event.stim_params['id'])[:,0]) & (events.type=='BIOMARKER')] = matched_events
         return events
-
-
-#
-# class VocalizationParser(BaseSys3LogParser,FRSessionLogParser):
-#     def __init__(self,protocol, subject, montage, experiment, session, files):
-#         super(VocalizationParser,self).__init__(protocol, subject, montage, experiment, session, files,
-#                                                 primary_log='event_log',allow_unparsed_events=True)
-#
-#         self._type_to_new_event = {
-#             'VOCALIZATION':self.event_default
-#         }
-#
-# class BiomarkerParser(BaseSys3LogParser,FRSessionLogParser):
-#     STIM_PARAMS_FIELD= 'msg_stub'
-#     BIOMARKER_FIELD = 'biomarker_value'
-#     ID_FIELD = 'hashtag'
-#     POSITION_FIELD ='buffer_name'
-#
-#     _STIM_FIELDS = BaseLogParser._STIM_FIELDS + (
-#             ('biomarker_value',-1.,'float64'),
-#             ('id','','S40'),
-#             ('pre_stim',True,'bool')
-#         )
-#
-#     def __init__(self,*args):
-#         super(BiomarkerParser,self).__init__(*args,primary_log='event_log',allow_unparsed_events=True)
-#
-#         self._type_to_new_event = {
-#             'BIOMARKER':self.event_biomarker
-#         }
-#
-#     def event_biomarker(self,event_json):
-#         event= self.event_default(event_json)
-#         json_stim_params=event_json[self.STIM_PARAMS_FIELD]
-#         event.stim_params.biomarker_value = json_stim_params[self.BIOMARKER_FIELD]
-#         event.stim_params.id = json_stim_params[self.ID_FIELD]
-#         event.stim_params.pre_stim = 'post' not in json_stim_params[self.POSITION_FIELD]
-#         return event
-#
-#
-
-
-
-
-
-
-if __name__ == '__main__':
-    import glob
-    import os
-    from viewers.view_recarray import pprint_rec as ppr
-    # d = '/Users/iped/event_creation/tests/test_input/R9999X/behavioral/PS2/session_37/host_pc/123_45_6788'
-    # event_log = os.path.join(d, 'event_log.json')
-    # conf = os.path.join(d, 'config_files', 'SubjID_TwoStimChannels.csv')
-    #
-    # s3lp = System3LogParser([event_log], [conf])
-    # ppr(s3lp.stim_events.view(np.recarray).stim_params[:,0])
-    event_log = ['/Users/leond/Downloads/event_log.json.txt']
-    # ann_files = glob.glob('/Volumes/rhino_root/data/eeg/R1286J/behavioral/catFR3/session_0/*.ann')
-    # VP = VocalizationParser('r1','R1999X','0.0','FR5','0',{'event_log':event_log, 'annotations':ann_files})
-    # events = VP.parse()
-    conf = '/Volumes/rhino_root/data/eeg/R1286J/behavioral/catFR3/session_0/host_pc/20170314_165146/config_files/R1286J_14MAR2017L1M1STIM.csv'
-    parser = Sys3EventsParser('r1','R1999X','0.0','FR5','0',{'event_log':event_log,'electrode_config':[conf]}) #'annotations':ann_files})
-    events = parser.parse()
-    pass
